+++ src/ui/dialogs/configdialog.cpp
@@ -261,6 +261,16 @@ void ConfigDialog::createChartTab()
     hbox->addWidget(m_markerSize);
     layout->addLayout(hbox);
 
+    m_lineWidth = new QDoubleSpinBox;
+    m_lineWidth->setMinimum(0);
+    m_lineWidth->setMaximum(30);
+    m_lineWidth->setValue(qApp->instance()->property("lineWidth").toDouble());
+
+    hbox = new QHBoxLayout;
+    hbox->addWidget(new QLabel(tr("Define line width for exported charts:")));
+    hbox->addWidget(m_lineWidth);
+    layout->addLayout(hbox);
+
     m_XScale = new QSpinBox;
     m_XScale->setMinimum(0);
     m_XScale->setMaximum(1e5);
@@ -348,6 +358,7 @@ void ConfigDialog::accept()
     qApp->instance()->setProperty("save_on_exit", m_save_on_exit->isChecked());
     qApp->instance()->setProperty("tooltips", m_tooltips->isChecked());
     qApp->instance()->setProperty("markerSize", m_markerSize->value());
+    qApp->instance()->setProperty("lineWidth", m_lineWidth->value());
     qApp->instance()->setProperty("xSize", m_XScale->value());
     qApp->instance()->setProperty("ySize", m_YScale->value());
     qApp->instance()->setProperty("transparentChart", m_transparentChart->isChecked());